+++ mountpoint-s3/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## Unreleased (v1.22.1)
 
 ## v1.22.0 (January 22, 2026)