+++ mountpoint-s3/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased (v1.22.1)
 
+* Upgrade cargo dependencies.
+
 ## v1.22.0 (January 22, 2026)
 
 ### Breaking changes