+++ mountpoint-s3/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased (v1.22.1)
 
+* Fix a race condition where concurrent operations after closing a truncated file could result in I/O errors on subsequent reads. The issue was introduced in v1.22.0.
+  ([#1781](https://github.com/awslabs/mountpoint-s3/pull/1781))
 * Upgrade cargo dependencies.
 
 ## v1.22.0 (January 22, 2026)